Modified the application to use Flask-JWT-Extended instead of Flask-JWT.

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 class Item(Resource):
@@ -16,6 +16,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': f"An item with name '{name}' already exists."}, 400
@@ -27,8 +28,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
-        
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -49,6 +54,13 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        items = ItemModel.query.all()
-        return {'items': [item.json() for item in items]}, 200
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in.'
+        }, 200
